install.py: do not try to expand a non-link

+++ install.py
@@ -4,15 +4,15 @@ import os
 import os.path
 import shutil
 
-pwd = os.path.dirname(os.path.realpath(__file__))
+PWD = os.path.dirname(os.path.realpath(__file__))
 
 
 def symlink_file(file_name, symlink_name):
     full_symlink_name = os.path.expanduser(symlink_name)
-    full_config_name = os.path.realpath(pwd + "/" + file_name)
+    full_config_name = os.path.realpath(PWD + "/" + file_name)
 
     if os.path.exists(full_symlink_name):
-        if not os.path.exists(os.readlink(full_symlink_name)):
+        if os.path.islink(full_symlink_name) and not os.path.exists(os.readlink(full_symlink_name)):
             os.remove(full_symlink_name)
         elif os.path.samefile(full_symlink_name, full_config_name):
             return
